♻️ Refactored Window initialization - resolves #77
